+++ src/main/java/com/hxk/community/controller/QuestionController.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -29,7 +28,7 @@ public class QuestionController {
     展示问题详情
      */
     @GetMapping("/question/{id}")
-    public String goQuestion(@PathVariable(name ="id" )Integer id,
+    public String goQuestion(@PathVariable(name ="id" )Long id,
                              HttpServletRequest httpServletRequest,
                              Model model){
         //获取当前登录用户
